requirements.txt
Django==3.2.6
djangorestframework==3.12.4
django-extensions==3.1.3
django-heroku==0.3.1
gunicorn==20.1.0
djangorestframework_api_key==2.0.0
